[#S05P21A506-52] chore: Entity 수정

+++ backend/src/main/java/com/a506/mirinae/domain/donation/DonationReq.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Getter
 public class DonationReq {
     Long fundingId;
-    Long amount;
+    Double amount;
 
     public Donation toEntity(User user, Funding funding, String tx_id) {
         return Donation.builder()
